RTree

Description

Hilbert R-Tree

Programacin Funcional Avanzada (CI4251)

1 2012-06-17

Johan Gonzlez 07-40979

Andreina Garca 08-10406

Este mdulo implanta una estructura de datos que almacena y consulta datos geomtricos de rectngulos, dicha estructura es un R-Tree de Hilbert.

Tipos exportados.

data Rectangle

Constructors

R

Fields

ul :: Point

Vertice superior izquierdo (X0,Y1)

ll :: Point

Vertice inferior izquierdo (X0,Y0)

lr :: Point

Vertice inferior derecho (X1,Y0)

ur :: Point

Vertice superior derecho (X1,Y1)

data RTree

Constructors

Branch
Rama del rbol

Fields

hv :: HV

mbr :: Rectangle

childs :: Set RTree

Leaf
Hoja del rbol

Fields

hv :: HV

mbr :: Rectangle

rects :: Set Rectangle

Empty

rbol vacio

type Point

Arguments

= (Int, Int)
(X,Y)

Funciones sobre Rectangulos

Constructores

makeRect4 :: [Int] -> Rectangle

makeRect4 permite construir un rectngulo a partir de una lista de 8 enteros.

makeRect2 :: Point -> Point -> Rectangle

makeRect2 permite construir un rectngulo a partir de 2 puntos en el espacio.

Comparacion

intersects :: Rectangle -> Rectangle -> Bool

intersects determina si dos rectngulos se intersectan.

orderHV :: Rectangle -> Rectangle -> Ordering

orderHV establece la comparacin de dos rectangulos segun su nmero de Hilbert.

Funciones sobre el Arbol

Constructores.

fromList :: [Rectangle] -> RTree

fromList construye un rbol a partir de una lista de rectngulos.

fromList' :: Foldable t => t Rectangle -> RTree

fromList' contruye un rbol a partir de un conjunto de rectngulos utilizando la operacin de insert.

funciones basicas

insert :: RTree -> Rectangle -> Either String RTree

insert inserta un rectngulo en un rbol.

Si el rectngulo ya se encuentra en el rbol se reporta un error de rectngulo duplicado.

Si no, se devulve el nuevo rbol.

delete :: RTree -> Rectangle -> Either String RTree

delete elimina un rectngulo de un rbol.

Si el rectngulo se encuentra en el rbol se retorne un nuevo arbol que no contiene al rectngulo.

En caso contrario reporta un error de rectngulo no encontrado.

search :: RTree -> Rectangle -> Maybe [Rectangle]

search busca todos los rectngulos que se solapen con un rectngulo dado.

Si no existen solapamientos con el rectngulo suministrado no se devuelve nada.

Si existen se devuelve una lista de todos los rectngulos solapados.

Pruebas

test :: IO ()

test ejecuta todas las pruebas de quickcheck.